refactor(common.py): Added git.py module that contains git utilities
