+++ sdcm/db_stats.py
@@ -31,7 +31,8 @@ import requests
 
 from sdcm.es import ES
 from sdcm.test_config import TestConfig
-from sdcm.utils.common import normalize_ipv6_url, get_git_commit_id
+from sdcm.utils.common import normalize_ipv6_url
+from sdcm.utils.git import get_git_commit_id
 from sdcm.utils.decorators import retrying
 from sdcm.sct_events.system import ElasticsearchEvent
 from utils.build_system import get_job_name, get_job_url
